Move Unassigned Drawer unit tests to UnassignedDrawer.test.tsx
